+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+next version
+------------
+
+* Move eta_shrinkage results method to modeling.calculate_eta_shrinkage
+
 0.29.0 (2021-08-25)
 -------------------